Rename some gene details for clarity; tidy up logic for location.jsx. (Markup needs to be changed from inappropriate <dl> stuff that we are using basically because default bootstrap styling is convenient.)
