+++ scripts/components/result/details/location/dallianceBrowser.jsx
